Move showsLikes array to separate file
