+++ src/getShows.js
@@ -1,4 +1,4 @@
-const getShows = async (pg = 1) => {
+const getShows = async (pg = 0) => {
   const response = await fetch(`https://api.tvmaze.com/shows?page=${pg}`, {
     method: 'GET',
   });
